Moved generic API methods to new file
